+++ .github/workflows/run_tests_ci.yaml
@@ -4,7 +4,9 @@ on:
   push:
     branches: 
       - main
-      - 'add-test-cases'
+  pull_request:
+    branches:
+      - master
 
 jobs:
   headless-run:
