+++ cypress/support/helpers/testHelper.js
@@ -1,6 +1,6 @@
 const faker = require('faker')
 
-class TestHelper {
+export class TestHelper {
     /**
     * This generates a random string, can be used to generate names too
     * @param {number} options.maxLength - max length of string to generate
@@ -14,7 +14,4 @@ class TestHelper {
             return faker.random.alpha(8);    
         }
     }
-
-    
 }
-export default TestHelper
