+++ cypress/support/pages/index.js
@@ -1,0 +1,2 @@
+export * from "./dashboard";
+export * from "./login";
